+++ packages/gestalt-codemods/1.15.0-1.16.0/__testfixtures__/deprecatedMargin-box-replace.output.js
@@ -1,0 +1,13 @@
+// @flow
+import React from 'react';
+import { Box, Box as Boxie } from 'gestalt';
+
+export default function TestText() {
+  return (
+    <Boxie margin={1} fit>
+      <Box marginTop={1} fit />
+      <Box marginBottom={1} fit />
+      <Box marginStart={1} marginEnd={1} fit />
+    </Boxie>
+  );
+}